pyproject.toml
[project]
name = "migs-tab"
version = "0.1.0"
description = "Convert YouTube acoustic guitar tutorial videos into accurate tabs + extracted tips"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "yt-dlp>=2025.1.15",
    "demucs>=4.0.1",
    "torchcodec>=0.4",
    "basic-pitch[onnx]>=0.4.0",
    "pretty_midi>=0.2.10",
    "librosa>=0.10.2",
    "numpy>=1.26,<2.0",
    "pydantic>=2.7",
    "Pillow>=10",
]

[project.scripts]
migs-tab = "migs_tab.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/migs_tab"]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "pre-commit>=3.7",
    "pytest>=9.0.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
# A pragmatic starter set — codestyle, pyflakes, isort, common bugbears,
# and modern-Python upgrades. Tighten over time.
select = ["E", "F", "W", "I", "UP", "B"]
ignore = [
    "E501",  # line length — ruff format owns this
    "B008",  # function call in default argument (typer uses this idiomatically)
]

[tool.ruff.lint.per-file-ignores]
"src/migs_tab/cli.py" = ["B008"]

